se hace el fetch del useProducts en la orden, se limpia codigo en usePedido, se cambia nombre a ingles, y se ajustan las rutas que afecto el cambio
