+++ frontend/components/Hooks/useProducts.ts
@@ -6,16 +6,14 @@ import { URI_PRODUCT } from "../URI/URI";
 import { editProduct } from "../../api/products";
 import { ProductForm } from "../Interfaces/IProducts";
 import { useAuth } from "@/app/context/authContext";
-import { Ingredient } from "../../../backend/src/Ingredient/ingredient.entity";
 import { FormTypeProduct, TypeProduct } from "../Enums/view-products";
 
-export const useProductos = () => {
+export const useProducts = () => {
   const { getAccessToken } = useAuth();
   const [token, setToken] = useState<string | null>(null);
   const {
     products,
     setProducts,
-    addProduct,
     removeProduct,
     updateProduct,
     connectWebSocket,
